Ability to set rules for routing to the application
Added the ability to set rules for routing to an application, so that any application itself can set the rules it needs and
we did not have to do this in the anonymous function of the container. It also allows changing routers and processing rules in the adapter of a specific router.